application/back/requirements.txt
fastapi==0.115.4
ultralytics==8.3.27
uvicorn==0.32.0
python-multipart==0.0.17
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
pillow==10.4.0
easyocr==1.7.2
paddleocr==2.9.1
paddlepaddle==2.6.2
onnx==1.17.0
onnxruntime==1.20.0